+++ src/models/post.js
@@ -243,7 +243,9 @@ var Post = function(filename, markdown, meta, config) {
    */
   internal.galleryHtml = function(html) {
     return html
+      .replace(/<p>(\s*(?:<img[^>]+>\s*){2,})<\/p>/g, '<div class="gallery">$1</div>')
       .replace(/(<img[^>]+src="([^"]+)(?:\-\d+x\d+)(\.(?:jpg|png|gif))"[^>]*>)/g, '<a href="$2$3" class="image">$1</a>')
+      .replace(/(<a[^>]+)(><img[^>]+alt=")(.+?)(")/g, '$1 title="$3"$2$3$4')
     ;
   };